+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/PlaylistPanel.java
@@ -7,7 +7,7 @@ import is.codion.framework.demos.chinook.domain.Chinook.PlaylistTrack;
 import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityPanel;
 
-import static is.codion.swing.framework.ui.TabbedPanelLayout.splitPaneResizeWeight;
+import static is.codion.swing.framework.ui.TabbedDetailLayout.splitPaneResizeWeight;
 
 public final class PlaylistPanel extends EntityPanel {
 
@@ -15,7 +15,7 @@ public final class PlaylistPanel extends EntityPanel {
     super(playlistModel,
             new PlaylistEditPanel(playlistModel.editModel()),
             new PlaylistTablePanel(playlistModel.tableModel()),
-            config -> config.panelLayout(splitPaneResizeWeight(0.25)));
+            config -> config.detailLayout(splitPaneResizeWeight(0.25)));
 
     SwingEntityModel playlistTrackModel = playlistModel.detailModel(PlaylistTrack.TYPE);
     EntityPanel playlistTrackPanel = new EntityPanel(playlistTrackModel,