+++ changelog.md
@@ -3,6 +3,8 @@ Codion Change Log
 
 ## 0.17.33-SNAPSHOT
 - EntityPanel, initialization refactored.
+### is.codion.swing.framework.ui
+- EntityPanel, initialization refactored further, PanelLayout renamed DetailLayout, now only handles laying out entity panels containing one or more detail panels.
 
 ## 0.17.32
 ### is.codion.common.core